requirements.txt
httpx>=0.27.0
requests>=2.31.0
requests-oauthlib>=2.0.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
freezegun>=1.4.0
duckduckgo-search>=6.0.0
